requirements.txt
# Core Framework
crewai>=0.80.0
crewai-tools>=0.12.0
crewai[google-genai]
# LLM Support
google-generativeai>=0.8.0
# Backend
Flask>=2.0.0
Flask-Cors>=3.0.0

# Environment Management
python-dotenv>=0.19.0
